complete changes + added exception handling test to algo_test
