+++ setup.py
@@ -4,7 +4,7 @@ from setuptools import setup
 
 setup(
     name='algorithmia',
-    version='1.7.1',
+    version='1.7.2',
     description='Algorithmia Python Client',
     long_description='Algorithmia Python Client is a client library for accessing Algorithmia from python code. This library also gets bundled with any Python algorithms in Algorithmia.',
     url='http://github.com/algorithmiaio/algorithmia-python',
